test(arc-layer): colocate antialiasing render coverage
